@@ -13,48 +13,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@capacitor/android": "~1.4.0",
-    "@capacitor/cli": "~1.4.0",
-    "@capacitor/core": "~1.4.0",
-    "@capacitor/ios": "~1.4.0",
-    "@ionic-enterprise/identity-vault": "~3.5.0",
-    "@ionic/angular": "~4.11.7",
+    "@angular/common": "~9.0.2",
+    "@angular/core": "~9.0.2",
+    "@angular/forms": "~9.0.2",
+    "@angular/platform-browser": "~9.0.2",
+    "@angular/platform-browser-dynamic": "~9.0.2",
+    "@angular/router": "~9.0.2",
+    "@capacitor/android": "~1.5.0",
+    "@capacitor/cli": "~1.5.0",
+    "@capacitor/core": "~1.5.0",
+    "@capacitor/ios": "~1.5.0",
+    "@ionic-enterprise/identity-vault": "~3.6.1",
+    "@ionic/angular": "~5.0.1",
     "@ionic/storage": "~2.2.0",
-    "core-js": "~3.5.0",
-    "rxjs": "~6.5.3",
+    "core-js": "~3.6.4",
+    "rxjs": "~6.5.4",
+    "tslib": "~1.11.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.803.20",
-    "@angular-devkit/build-angular": "~0.803.20",
-    "@angular-devkit/core": "~8.3.20",
-    "@angular-devkit/schematics": "~8.3.20",
-    "@angular/cli": "~8.3.20",
-    "@angular/compiler": "~8.2.14",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@ionic/angular-toolkit": "~2.1.1",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/architect": "~0.900.3",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "@angular-devkit/core": "~9.0.3",
+    "@angular-devkit/schematics": "~9.0.3",
+    "@angular/cli": "~9.0.3",
+    "@angular/compiler": "~9.0.2",
+    "@angular/compiler-cli": "~9.0.2",
+    "@angular/language-service": "~9.0.2",
+    "@ionic/angular-toolkit": "~2.2.0",
+    "@types/jasmine": "~3.5.7",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.12.18",
-    "codelyzer": "~5.2.0",
+    "@types/node": "~13.7.4",
+    "codelyzer": "~5.2.1",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.1.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "~1.4.2",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.5.4",
-    "tslint": "~5.20.1",
-    "typescript": "~3.5.3"
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "~1.5.2",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.6.2",
+    "tslint": "~6.0.0",
+    "typescript": "~3.7.5"
   },
   "description": "An Ionic project"
 }
