@@ -4,7 +4,10 @@ import { Routes, RouterModule } from '@angular/router';
 const routes: Routes = [
   { path: '', loadChildren: () => import('./tabs/tabs.module').then(m => m.TabsPageModule) },
   { path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginPageModule) },
-  { path: 'edit-tea-category', loadChildren: () => import('./edit-tea-category/edit-tea-category.module').then(m => m.EditTeaCategoryPageModule) }
+  {
+    path: 'edit-tea-category',
+    loadChildren: () => import('./edit-tea-category/edit-tea-category.module').then(m => m.EditTeaCategoryPageModule)
+  }
 ];
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
