@@ -18,14 +18,12 @@ export class AppComponent {
     const { SplashScreen, StatusBar } = Plugins;
     await this.identity.ready();
     this.identity.get();
-    try {
+    if (this.platform.is('hybrid')) {
       await SplashScreen.hide();
       await StatusBar.setStyle({ style: StatusBarStyle.Light });
       if (this.platform.is('android')) {
         StatusBar.setBackgroundColor({ color: '#3171e0' });
       }
-    } catch (err) {
-      console.log('This is normal in a browser', err);
     }
   }
 }
