@@ -24,7 +24,7 @@ describe('AuthenticationService', () => {
       ]
     });
 
-    httpTestingController = TestBed.get(HttpTestingController);
+    httpTestingController = TestBed.inject(HttpTestingController);
   });
 
   beforeEach(inject(
